Address, card, category service exceptions implemented.

+++ src/main/java/com/unicarioca/projeto_poo/service/AddressService.java
@@ -2,7 +2,10 @@ package com.unicarioca.projeto_poo.service;
 
 import com.unicarioca.projeto_poo.domain.address.Address;
 import com.unicarioca.projeto_poo.domain.address.AddressRequestDTO;
+import com.unicarioca.projeto_poo.domain.card.Card;
 import com.unicarioca.projeto_poo.domain.client.Client;
+import com.unicarioca.projeto_poo.exception.ElementNotFoundException;
+import com.unicarioca.projeto_poo.exception.ExistingElementException;
 import com.unicarioca.projeto_poo.repository.AddressRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,6 +26,9 @@ public class AddressService {
     public Address createAddress(AddressRequestDTO addressDTO){
         Address address = new Address();
         Client client = clientService.getClientById(addressDTO.id_client());
+        if (verifyExistingAddress(addressDTO.street())){
+            throw new ExistingElementException();
+        }
 
         address.setStreet(addressDTO.street());
         address.setComplement(addressDTO.complement());
@@ -36,6 +42,9 @@ public class AddressService {
 
     public Address updateAddress(UUID idAddress, AddressRequestDTO addressDTO){
         Address address = addressRepository.findById(idAddress).get();
+        if (verifyExistingAddress(addressDTO.street())){
+            throw new ExistingElementException();
+        }
 
         address.setStreet(addressDTO.street());
         address.setNumber(addressDTO.number());
@@ -46,8 +55,18 @@ public class AddressService {
         return address;
     }
 
-    public void deleteAddress(UUID id){
-        addressRepository.deleteById(id);
+    public void deleteAddress(UUID id) {
+        if(addressRepository.existsById(id)){
+            addressRepository.deleteById(id);
+        } else {
+            throw new ElementNotFoundException();
+        }
+    }
+
+    public boolean verifyExistingAddress(String street) {
+        Address address = addressRepository.findAddressByStreet(street);
+
+        return address != null;
     }
 
 }
